refactor: replace GST Purchase Register with GST Purchase Register Beta
