+++ india_compliance/gst_india/report/gst_purchase_register/gst_purchase_register.json
@@ -1,15 +1,15 @@
 {
  "add_total_row": 1,
  "columns": [],
- "creation": "2017-06-22 11:06:40.836073",
- "disable_prepared_report": 0,
+ "creation": "2024-08-26 20:42:05.592632",
  "disabled": 0,
  "docstatus": 0,
  "doctype": "Report",
  "filters": [],
  "idx": 0,
  "is_standard": "Yes",
- "modified": "2022-03-01 15:45:47.658452",
+ "letterhead": null,
+ "modified": "2024-08-26 20:42:05.592632",
  "modified_by": "Administrator",
  "module": "GST India",
  "name": "GST Purchase Register",
@@ -20,10 +20,16 @@
  "report_type": "Script Report",
  "roles": [
   {
-   "role": "Accounts Manager"
+   "role": "Purchase User"
   },
   {
    "role": "Accounts User"
+  },
+  {
+   "role": "Auditor"
+  },
+  {
+   "role": "Accounts Manager"
   }
  ]
 }
