--- india_compliance/gst_india/report/gst_purchase_register_beta/gst_purchase_register_beta.json
@@ -1,35 +1,0 @@
-{
- "add_total_row": 1,
- "columns": [],
- "creation": "2024-08-26 20:42:05.592632",
- "disabled": 0,
- "docstatus": 0,
- "doctype": "Report",
- "filters": [],
- "idx": 0,
- "is_standard": "Yes",
- "letterhead": null,
- "modified": "2024-08-26 20:42:05.592632",
- "modified_by": "Administrator",
- "module": "GST India",
- "name": "GST Purchase Register Beta",
- "owner": "Administrator",
- "prepared_report": 0,
- "ref_doctype": "Purchase Invoice",
- "report_name": "GST Purchase Register Beta",
- "report_type": "Script Report",
- "roles": [
-  {
-   "role": "Purchase User"
-  },
-  {
-   "role": "Accounts User"
-  },
-  {
-   "role": "Auditor"
-  },
-  {
-   "role": "Accounts Manager"
-  }
- ]
-}
